refactor(discovery): share parseSeeds via runtime
