+++ src/core/discovery-modes/modes/runtime.ts
@@ -40,3 +40,24 @@ export function getProviderPath(request: DiscoveryModeRequest): string[] {
 export function normalizeDiscoveryName(name: string): string {
   return name.trim().toLowerCase()
 }
+
+export type SeedArtist = { name: string; mbid?: string }
+
+export function parseSeeds(raw: unknown): SeedArtist[] {
+  const items = Array.isArray(raw)
+    ? raw
+    : typeof raw === 'string'
+      ? raw.split(',').map((s) => s.trim())
+      : []
+  const seeds: SeedArtist[] = []
+  for (const item of items) {
+    if (typeof item === 'string') {
+      if (item.trim()) seeds.push({ name: item.trim() })
+    } else if (item && typeof item === 'object' && 'name' in item) {
+      const rec = item as Record<string, unknown>
+      const name = String(rec.name ?? '').trim()
+      if (name) seeds.push({ name, mbid: typeof rec.mbid === 'string' ? rec.mbid : undefined })
+    }
+  }
+  return seeds
+}
